Fix the tuple table slot description
When the Result plan node is created it is getting dummy TLEs in place
of the missing attributes, and those dummy attributes are getting into
the tuple table slot description. The mismatch is not a problem, except
if the missing attributes are retrieved from the DocDB. In this case the
descriptor should be fixed.